// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  Admin: 'Admin',
  VP4HR: 'VP4HR',
  FullMember: 'FullMember'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const MemberStatus = {
  Observer: 'Observer',
  Baby: 'Baby',
  Full: 'Full',
  Alumni: 'Alumni'
} as const

export type MemberStatus = (typeof MemberStatus)[keyof typeof MemberStatus]


export const MemberState = {
  Active: 'Active',
  Inactive: 'Inactive'
} as const

export type MemberState = (typeof MemberState)[keyof typeof MemberState]


export const Gender = {
  Male: 'Male',
  Female: 'Female'
} as const

export type Gender = (typeof Gender)[keyof typeof Gender]


export const TeamType = {
  Coreteam: 'Coreteam',
  Project: 'Project',
  Team: 'Team'
} as const

export type TeamType = (typeof TeamType)[keyof typeof TeamType]


export const ActivityType = {
  Gathering: 'Gathering',
  SIT: 'SIT',
  LeisureEvent: 'LeisureEvent'
} as const

export type ActivityType = (typeof ActivityType)[keyof typeof ActivityType]


export const ApplicationResult = {
  Success: 'Success',
  Fail: 'Fail'
} as const

export type ApplicationResult = (typeof ApplicationResult)[keyof typeof ApplicationResult]